+++ StreamChatSwiftUITests/Tests/ChatChannelList/MoreChannelActionsView_Tests.swift
@@ -29,6 +29,6 @@ class MoreChannelActionsView_Tests: StreamChatTestCase {
         .applyDefaultSize()
 
         // Then
-        assertSnapshot(matching: view, as: .image)
+        assertSnapshot(matching: view, as: .image(perceptualPrecision: 0.98))
     }
 }