+++ CHANGELOG.md
@@ -3,6 +3,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### 🔄 Changed
+
+# [4.36.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.36.0)
+_August 31, 2023_
+
 ### ✅ Added
 - Add XCPrivacy manifest [#352](https://github.com/GetStream/stream-chat-swift/pull/352)